requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-decouple==3.8
requests==2.31.0
PyJWT==2.8.0